refactor cli tx
